backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
PyJWT==2.8.0
Werkzeug==2.3.7
bcrypt==4.1.2
qrcode[pil]==7.4.2
Pillow==10.2.0
numpy>=1.26.0
